Move `editable_tree::dag::DAG` to `editable_tree::DAG`
